+++ odoo/local-src/specific_sale/__manifest__.py
@@ -5,7 +5,13 @@
 {
     "name": "RocTool specific sale module",
     "version": "10.0.1.0.0",
-    "depends": ['specific_crm', 'sale', 'sale_crm', 'website_sale_options'],
+    "depends": [
+        'specific_crm',
+        'sale',
+        'sale_crm',
+        'website_sale_options',
+        'mrp',
+    ],
     "author": "Camptocamp,Odoo Community Association (OCA)",
     "website": "http://www.camptocamp.com",
     "license": "GPL-3 or any later version",
@@ -13,6 +19,8 @@
     "data": [
         'views/product_views.xml',
         'views/sale_order_crm.xml',
+        'views/project_task.xml',
+        'views/mrp_bom.xml',
         'data/res_groups_data.xml',
     ],
     'installable': True,